COLORS! Added code improvements Resolved a newline bug in name

+++ backup manager.py
@@ -1,5 +1,6 @@
 from os import system, path
 from time import sleep
+import console as term
 # Incluir el instalador en el menu de inicio
 def cmd(c):
 	_ = system(c)
@@ -16,9 +17,7 @@ while True:
 				_ = open(USB+":/.USBignore")
 				_.close()
 				bname = file.read()
-				file.close()
-				if not USB in backed:
-					if ('"'  not in bname and
+				bnameCheck=('"'  not in bname and
 						'\\' not in bname and
 						'\n' not in bname and
 						'/'  not in bname and
@@ -28,16 +27,18 @@ while True:
 						'<'  not in bname and
 						'|'  not in bname and
 						'?'  not in bname and
-						len( bname ) != 0 ):
-						backed += USB
-						print('SAVING...',end='')
-						sleep(0.1)
-						cmd('rmdir /S /Q "%HOMEDRIVE%%HOMEPATH%/USB-backup/'+bname+'" >> %HOMEDRIVE%%HOMEPATH%/USB-backup/log.txt')
-						cmd('xcopy "'+USB+':/" "%HOMEDRIVE%%HOMEPATH%/USB-backup/'+bname+'" /S /E /I /EXCLUDE:'+USB+':\\.USBignore >> %HOMEDRIVE%%HOMEPATH%/USB-backup/log.txt')
-					else:
-						print(' INVALID ',end='')
+						len( bname ) != 0 )
+				file.close()
+				if USB not in backed and bnameCheck:
+					backed += USB
+					print('SAVING...',end='')
+					sleep(0.1)
+					cmd('rmdir /S /Q "%HOMEDRIVE%%HOMEPATH%/USB-backup/'+bname+'" >> %HOMEDRIVE%%HOMEPATH%/USB-backup/log.txt')
+					cmd('xcopy "'+USB+':/" "%HOMEDRIVE%%HOMEPATH%/USB-backup/'+bname+'" /S /E /I /EXCLUDE:'+USB+':\\.USBignore >> %HOMEDRIVE%%HOMEPATH%/USB-backup/log.txt')
+				elif not bnameCheck:
+					print('\x1b[101;1m INVALID \x1b[0m',end='')
 				else:
-					print('BACKED UP',end='')
+					print('\x1b[92;1mBACKED UP\x1b[0m',end='')
 					# cmd("start copyscript.bat "+USB+" "+bname)
 			except OSError:
 				bname=""
@@ -47,7 +48,8 @@ while True:
 				else:
 					print('NOT FOUND',end='')
 			finally:
-				print(']'+'\t'+('Name: ')*(len(bname)>0)+bname)
+				print(']'+'\t'+('Name: ')*(len(bname)>0)+bname.replace('\n','<NEWLINE>'))
 		sleep(5)
 	except KeyboardInterrupt:
 		cmd("pause")
+		exit()
